requirements.txt
PyGithub==1.35
PyYAML==3.12
pygit2==0.26.1
